+++ src/main/java/com/nhnacademy/store99/bookstore/common/response/CommonResponse.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.store99.bookstore.common.dto;
+package com.nhnacademy.store99.bookstore.common.response;
 
 import lombok.Builder;
 import lombok.Getter;
