+++ src/main/java/com/nhnacademy/store99/bookstore/secure_key_manager/exception/SecureKeyMangerException.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.store99.bookstore.common.exception;
+package com.nhnacademy.store99.bookstore.secure_key_manager.exception;
 
 public class SecureKeyMangerException extends RuntimeException {
     public static final String MESSAGE = "Secure Key Manager 오류 발생 : ";
