+++ src/main/java/com/nhnacademy/store99/bookstore/secure_key_manager/response/SecretBodyResponse.java
@@ -1,0 +1,12 @@
+package com.nhnacademy.store99.bookstore.secure_key_manager.response;
+
+import lombok.Setter;
+
+@Setter
+public class SecretBodyResponse {
+    private String secret;
+
+    public String getSecret() {
+        return secret;
+    }
+}
